requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tensorflow==2.13.0
keras==2.13.1
jupyter==1.0.0
notebook==6.5.4
pyyaml==6.0.1
pytest==7.4.0
black==23.7.0
pylint==2.17.5
tqdm==4.65.0
joblib==1.3.1
plotly==5.15.0
shap==0.42.1
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.1
